Removed non-RESTful routes on groups_controller.
Made a dedicated controller for those routes called MembershipsController
Modified it to follow RESTful verbs
Edited helpers and routes also.
