+++ app/controllers/groups_controller.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-# rubocop:disable ClassLength
 class GroupsController < ApplicationController
   before_action :authenticate_user!
   before_action :set_group, only: %i[show edit update destroy]
@@ -71,45 +70,7 @@ class GroupsController < ApplicationController
       end
     end
   end
-  # rubocop:enable MethodLength
 
-  def join
-    @group_member = GroupMember.create!(group_member_params)
-    @group = @group_member.group
-    GroupNotifier.new(@group, 'new_group_member', current_user)
-                 .send_notifications_to(@group.leaders)
-
-    flash[:notice] = t('groups.join_success')
-    redirect_to_group
-  end
-
-  # rubocop:disable MethodLength
-  def leave
-    member_id = params[:memberid] || current_user.id
-    group_member = GroupMember.find_by(
-      userid: member_id,
-      groupid: params[:groupid]
-    )
-    group = group_member.group
-
-    # Cannot leave When you are the only leader
-    if group.leader_ids == [member_id]
-      flash[:alert] = t('groups.leave.error')
-    else
-      group_member.destroy
-
-      flash[:notice] = if member_id == current_user.id
-                         t('groups.leave.success', group: group.name)
-                       else
-                         t(
-                           'groups.leave.remove_member_success',
-                           user: group_member.user.name,
-                           group: group.name
-                         )
-                       end
-    end
-    redirect_to_index
-  end
   # rubocop:enable MethodLength
 
   # DELETE /groups/1
@@ -131,10 +92,6 @@ class GroupsController < ApplicationController
 
   def group_params
     params.require(:group).permit(:name, :description)
-  end
-
-  def group_member_params
-    params.permit(:groupid).merge(userid: current_user.id, leader: false)
   end
 
   def update_leaders
@@ -172,4 +129,3 @@ class GroupsController < ApplicationController
     end
   end
 end
-# rubocop:enable ClassLength
