+++ spec/controllers/groups/memberships_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Groups::MembershipsController, type: :controller do
+
+end
